Module Lib.Make

Configuration builder: stage 1

Parameters

module _ : S

Signature

val register : ?init:DSL.job DSL.impl list -> ?src:[ `Auto | `None | `Some of string ] -> string -> DSL.job DSL.impl list -> unit

register name jobs registers the application named by name which will execute the given jobs. Same optional arguments as DSL.main.

init is the list of job to execute before anything else (such as command-line argument parsing, log reporter setup, etc.). The jobs are always executed in the sequence specified by the caller.