Module Mirage_key.Alias

Alias allows to define virtual keys in terms of other keys at configuration time only.

Alias

type 'a t = 'a Functoria_key.Alias.t

The type for key alias.

val add : 'b key -> ('a -> 'b option) -> 'a t -> 'a t

add k f a set a as an alias for the key k: setting a on the command-line will set k to f applied to a's value. If f returns None, no value is set.

val flag : Arg.info -> bool t

flag is similar to Arg.flag but defines configure-only command-line flag alias. Set stage to `Configure.